RavenDB-26053 Split components
